feat(api): render Markdown from OxContentOptions (#877)
