+++ npm/vite-plugin-ox-content/src/public-exports.test.ts
@@ -24,6 +24,8 @@ describe("public export surface", () => {
       "generateMarkdown",
       "isMarkdownFilePath",
       "oxContent",
+      "createMarkdownProcessor",
+      "renderMarkdown",
       "resolveCollectionsOptions",
       "renderHtmlToReactCreateElement",
       "renderHtmlToVueH",
